Fill Names of IDs

The Facebook data contains variables with IDs, where it can be useful to have the name of the associated ID. This function creates variables that contain the the name. In cases where there are multiple IDs, the names are separated by a semicolon. For example, if there is an interest variable with the observation: "6003349442621,6003020834693", an interest_name variable with "Entertainment; Music" would be created.

fill_fb_id_names(fb_df, version, token)

Arguments

fb_df

Facebook dataframe, returned by query_fb_marketing_api()

version

Facebook Marketing API version; for example, "v14.0"

token

Facebook Marketing API token

Value

Dataframe with Facebook results, where additional variables are created that indicate the name of IDs

Examples

if (FALSE) {
fb_df <- query_fb_marketing_api(location_unit_type = "coordinates",
                                lat_lon = c(38.90, -77.01),
                                radius = 10,
                                radius_unit = "kilometer",
                                interest = list(c(6002839660079,
                                                  6002884511422),
                                                6002839660079,
                                                6002884511422),
                                behavior = list(c(6002714895372,
                                                  6002714898572),
                                                6002714895372,
                                                6002714898572),
                                version = VERSION, 
                                creation_act = CREATION_ACT, 
                                token = TOKEN)

fb_df <- fill_fb_id_names(fb_df,
                         version = VERSION,
                         token = TOKEN)
}