Query Facebook Marketing API

Location

Query Facebook Marketing API

Location

query_fb_marketing_api(
  location_unit_type,
  lat_lon = NULL,
  radius = NULL,
  radius_unit = NULL,
  location_keys = NULL,
  location_types = "home",
  locales = NULL,
  interests = NULL,
  behaviors = NULL,
  college_years = NULL,
  education_majors = NULL,
  education_schools = NULL,
  education_statuses = NULL,
  family_statuses = NULL,
  income = NULL,
  industries = NULL,
  work_positions = NULL,
  work_employers = NULL,
  excl_interests = NULL,
  excl_behaviors = NULL,
  excl_college_years = NULL,
  excl_education_majors = NULL,
  excl_education_schools = NULL,
  excl_education_statuses = NULL,
  excl_family_statuses = NULL,
  excl_income = NULL,
  excl_industries = NULL,
  excl_work_positions = NULL,
  excl_work_employers = NULL,
  relationship_statuses = NULL,
  life_events = NULL,
  user_os = NULL,
  wireless_carrier = NULL,
  gender = c(1, 2),
  age_min = 18,
  age_max = 65,
  flex_target = NULL,
  version,
  creation_act,
  token,
  sleep_time = 0.1,
  show_result = T,
  add_param_id_name_vars = F,
  add_query = F,
  add_query_hide_credentials = T
)

Arguments

location_unit_type

Either "coordinates" (for buffer around single point) or "country"

If location_unit_type = "coordinates"

lat_lon

Coordinates, c(lat, lon). For example, c(38.90, -77.01)

radius

Radius around coordinate

radius_unit

Unit for radius; either "kilometer" or "mile"

If location_unit_type = "country"

location_types

Either: (1) "home" (people whose stated location in Facebook profile "current city" is in an area, valided by IP), (2) "recent" (people whose recent location is in the selected area, determined by mobile device data), (3) "travel_in" (people whose most recent location is in selected area and more than 100 miles from stated current city), (4) c("home", "recent") (for people in either category)

locales

Locales ID.

education_statuses

Education status IDs. If multiple, uses OR condition; for example, education_statuses = c(9,10) will yeild those who report to have either a Master degree or professional degree. See education_statuses in the Advanced Targeting Documentation to see education status options.

family_statuses

Vector of family status IDs. If multiple, uses OR condition; for example, family_statuses = c(6023080302983, 6023005681983) targets users who are parents with preteens or parents with teenagers. Use get_fb_parameters(type = "demographics") to get dataframe with IDs and descriptions.

income

Vector of income IDs. If multiple, uses OR condition; for example, income = c(6107813553183, 6107813554583) targets users with a household income in the top 10%-25% or 25%-50% of ZIP codes (US). Use get_fb_parameters(type = "demographics") to get dataframe with IDs and descriptions.

industries

Vector of industries IDs. If multiple, uses OR condition; for example, industries = c(6008888980183, 6008888972183) targets users who work in sales or legal services. Use get_fb_parameters(type = "demographics") to get dataframe with IDs and descriptions.

relationship_statuses

Vector of relationship status IDs. If multiple, uses OR condition; for example, relationship_statuses = c(3,4) targets those who are married or engaged. See relationship_statuses in the Advanced Targeting Documentation to see relationship status ID options with descriptions.

life_events

Vector of life event IDs. If multiple, uses OR condition; for example, life_events = c(6005149512172, 6005149512172) targets users who recently moved or are in a new job. Use get_fb_parameters(type = "demographics") to get dataframe with IDs and descriptions.

user_os

User operating systems. If multiple, uses OR condition; for example user_os = ['iOS', 'Android'] targets those that use either an iOS or Android OS. See user_os in the Advanced Targeting Documentation for additional details.

wireless_carrier

Wireless carriet. If set to Wifi, then targets those connecting via a Wifi network. See wireless_carrier in the Advanced Targeting Documentation for additional details.

gender

Genders to target; 1 targets males and 2 targets females Default is both. See gender in the Basic Targeting Documentation.

age_min

Minimum age. Default is 18. See age_min in the Basic Targeting Documentation.

age_max

Maximum age. Default is 65. See age_max in the Basic Targeting Documentation.

Credentials

version

API version. e.g., "v14.0"

creation_act

Facebook creation act

token

Facebook API token

Scraping parameters

sleep_time

words

show_result

words

Return query text as variable in returned dataframe

add_query

If TRUE, add query text as variable in returned dataframe

add_query_hide_credentials

If TRUE (and add_query is TRUE), hide the creation_act and token from the query text returned in the dataframe

country_code

Country ISO2; for example, "US".

Other location??

behavior

Vector of behavior IDs. If multiple, uses OR condition; for example, behavior = c(6002714895372, 6008297697383) will target users who are either frequent travelers or returned from travels 2 weeks ago. Use get_fb_parameters(type = "behaviors") to get dataframe with IDs and descriptions.

interest

Vector of interest IDs. If multiple, uses OR condition; for example, interest = c(6003349442621, 6003139266461) will target users who are interested in either entertainment or movies. Use get_fb_parameters(type = "interests") to get dataframe with IDs and descriptions.

Value

Dataframe that includes (1) daily and monthly active users and (2) parameter values

Details

FOR LOOP, USE LISTS. BUT JUST CAN'T LIST ON LOCATION TYPE.

See also